Gold Store/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
arrow==1.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bs4==0.0.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
fastapi==0.112.0
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==13.7.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
